// src/app/models/commissionDTO.ts
export class commissionDTO {
    id: number;
    index: number;
    agentName: string;
    companyName: string;
    caseId: number;
    caseTypeId: number;
    clientName: string;
    referredOnDate: string;
    totalCommission: number;
    paid: number;
    balance: number;
    case:string;
    caseTypeName : string;

}